requirements.txt
Flask==3.1.0
Flask_Cors==5.0.0
python-dotenv==1.1.0
Requests==2.32.3
rich==13.9.4
SQLAlchemy==2.0.37
psycopg2_binary==2.9.10
